[opti]: move the block style widgets to style_widgets folder
